--- docs/principes/algoritmen/index.md
@@ -1,117 +1,0 @@
----
-title: "11. Pas algoritmen verantwoord toe"
-summary: Zorg voor transparante, eerlijke en verantwoorde toepassing van algoritmen en AI-systemen.
-relations:
-  - data
-  - veiligheid
-  - privacy
-  - gebruikersbehoefte
----
-
-# 11. Pas algoritmen verantwoord toe
-
-Algoritmen en AI-systemen hebben een steeds grotere impact op besluitvorming binnen de overheid. Een verantwoorde toepassing vereist transparantie, uitlegbaarheid, eerlijkheid en menselijk toezicht om de rechten van burgers te waarborgen, discriminatie te voorkomen en vertrouwen te behouden in digitale overheidsdiensten.
-
-## Waarom is het belangrijk?
-
-- **Rechtvaardigheid**: Voorkom discriminatie en oneerlijke behandeling van burgers
-- **Transparantie**: Maak duidelijk hoe beslissingen tot stand komen
-- **Verantwoording**: Zorg dat verantwoordelijkheid duidelijk is belegd
-- **Vertrouwen**: Behoud publiek vertrouwen in overheidsdiensten
-- **Naleving regelgeving**: Voldoe aan wet- en regelgeving zoals de AI-verordening
-- **Kwaliteit**: Verbeter besluitvorming door goede algoritmegovernance
-- **Ethisch handelen**: Waarborg fundamentele rechten in geautomatiseerde processen
-- **Controleerbaarheid**: Stel burgers in staat om besluiten te begrijpen en aan te vechten
-
-Door algoritmen verantwoord toe te passen, zorgen overheidsorganisaties ervoor dat hun digitale systemen eerlijk, transparant en in lijn met publieke waarden functioneren, wat cruciaal is voor het behouden van vertrouwen in een steeds meer gedigitaliseerde overheid.
-
-## Hoe pas je het toe?
-
-<div class="direct-aan-de-slag">
-    <h3>Direct aan de slag</h3>
-
-    <div class="warning-banner" style="background-color: #fff3e0; padding: 0.5rem; border-left: 3px solid #ff9800; margin-bottom: 0.8rem;">
-        <strong>Work in Progress:</strong> De onderstaande functionaliteit is nog in ontwikkeling.
-    </div>
-
-    <div class="action-cards">
-        <div class="action-card">
-            <span class="wip-badge wip-badge-beschikbaar">beschikbaar</span>
-            <h4>Algoritmekader</h4>
-            <p>Richtlijnen voor verantwoord algoritmebeheer</p>
-            <a href="https://minbzk.github.io/Algoritmekader/" class="action-button" target="_blank">Verkennen</a>
-        </div>
-        <div class="action-card">
-            <span class="wip-badge wip-badge-beschikbaar">beschikbaar</span>
-            <h4>Beslishulp AI-verordening</h4>
-            <p>Bepaal direct welke voorschriften gelden</p>
-            <a href="https://ai-verordening-beslishulp.apps.digilab.network/" class="action-button" target="_blank">Starten</a>
-        </div>
-        <div class="action-card">
-            <span class="wip-badge wip-badge-beschikbaar">beschikbaar</span>
-            <h4>Algoritme Management Toolkit</h4>
-            <p>Register voor risicovolle algoritmen</p>
-            <a href="https://amt.prd.apps.digilab.network/" class="action-button" target="_blank">Registreren</a>
-        </div>
-        <div class="action-card">
-            <span class="wip-badge wip-badge-beschikbaar">beschikbaar</span>
-            <h4>LLM Benchmark</h4>
-            <p>Benchmark voor taalmodellen (LLMs)</p>
-            <a href="https://github.com/MinBZK/llm-benchmark" class="action-button" target="_blank">Verkennen</a>
-        </div>
-         <div class="action-card">
-            <span class="wip-badge wip-badge-beschikbaar">beschikbaar</span>
-            <h4>Algoritmeregister</h4>
-            <p>Registreer je (impactvolle) algoritme in het nationale register</p>
-            <a href="https://algoritmes.overheid.nl/nl" class="action-button" target="_blank">Verkennen</a>
-        </div>
-        <div class="action-card">
-            <span class="wip-badge wip-badge-beschikbaar">beschikbaar</span>
-            <h4>Unsupervised Biasdetection Tool</h4>
-            <p>Een statistisch hulpmiddel dat groepen identificeert waar een algoritme afwijkende prestaties laat zien</p>
-            <a href="https://algorithmaudit.eu/technical-tools/bdt/" class="action-button" target="_blank">Verkennen</a>
-        </div>
-    </div>
-</div>
-
-### Implementatiestappen
-
-- **Algoritme-impact-assessment**: Voer een risicoanalyse uit met het Algoritmekader
-- **Ontwerp met waarden**: Integreer ethische waarden in het ontwerp
-- **Datagovernance**: Zorg voor kwaliteit en rechtvaardig gebruik van data
-- **Documentatie**: Documenteer het algoritme, gebruik en beslissingen
-- **Testen op bias**: Controleer actief op vooroordelen in algoritmen
-- **Menselijk toezicht**: Implementeer betekenisvolle menselijke controle
-- **Periodieke evaluatie**: Evalueer algoritmen regelmatig
-- **Transparantierapportage**: Publiceer informatie over het gebruik in het Algoritmeregister
-- **Classificatie**: Bepaal het risiconiveau volgens de AI-verordening
-
-### Uitdagingen en oplossingen
-
-- **Complexiteit vs. uitlegbaarheid**: Kies voor uitlegbare algoritmen waar mogelijk
-- **Opsporen van bias**: Implementeer tools en methoden voor het opsporen van vooroordelen
-- **Balans automatisering en menselijk oordeel**: Ontwerp hybride besluitvormingsprocessen
-- **Kennisgebrek**: Investeer in training en bewustwording
-- **Bestaande systemen**: Ontwikkel een plan voor het beoordelen van bestaande algoritmen
-
-### Algoritmeregistratie en -verantwoording
-
-- Identificeer en classificeer algoritmen op risico
-- Registreer hoog-risico algoritmen in het Algoritme Meldpunt
-- Registreer algoritme in het [Algoritmeregister](https://algoritmes.overheid.nl/nl)
-- Documenteer doel, werking en waarborgen
-- Maak openbaar algoritmebeschrijvingen voor hoog-risico algoritmen
-- Evalueer en actualiseer informatie regelmatig
-
-## Gerelateerde hulpmiddelen
-
-- **[Rijksbrede Algoritmekader](https://minbzk.github.io/Algoritmekader/)**: Kader voor risicoanalyse, verantwoorde inzet en compliance
-- **[Impact Assessment Mensenrechten en Algoritmen (IAMA)](https://www.rijksoverheid.nl/documenten/rapporten/2021/02/25/impact-assessment-mensenrechten-en-algoritmes)**: Methodiek voor het beoordelen van mensenrechtenimpact
-- **[Code Goed Digitaal Openbaar Bestuur](https://www.digitaleoverheid.nl/overzicht-van-alle-onderwerpen/nieuwe-technologieen-data-en-ethiek/publieke-waarden/code-goed-digitaal-openbaar-bestuur/)**: Richtlijnen voor digitaal bestuur
-
-## Gerelateerde principes
-
-- [1. Stel gebruikersbehoeften vast](../gebruikersbehoeften/index.md)
-- [6. Maak veilige systemen](../veiligheid/index.md)
-- [7. Maak privacy integraal](../privacy/index.md)
-- [10. Maak beter gebruik van data](../data/index.md)
