--- docs/principes/index.md
@@ -1,98 +1,0 @@
----
-title: Principes
-summary: De principes van de Nederlandse Richtlijn Digitale Systemen
----
-
-# Principes van de NeRDS
-
-De Nederlandse Richtlijn Digitale Systemen (NeRDS) is gebaseerd op een set principes die richting geven aan het ontwerpen, ontwikkelen en implementeren van digitale systemen binnen de Nederlandse overheid.
-
-<div class="grid cards" markdown>
-
-- :material-account-search:{ .lg .middle } __1. [Stel gebruikersbehoeften vast](gebruikersbehoeften/index.md)__
-
-    ---
-
-    Begrijp je gebruikers en hun behoeften. Ontwikkel kennis over je gebruikers en wat dat betekent voor je technologieproject.
-
-- :material-human-wheelchair:{ .lg .middle } __2. [Zorg voor toegankelijkheid en inclusie](toegankelijkheid/index.md)__
-
-    ---
-
-    Zorg ervoor dat je technologie, infrastructuur en systemen toegankelijk en inclusief zijn voor alle gebruikers.
-
-- :material-source-branch:{ .lg .middle } __3. [Werk transparant en gebruik open source](open-source/index.md)__
-
-    ---
-
-    Publiceer je code en gebruik open source software om transparantie, flexibiliteit en verantwoording te verbeteren.
-
-- :material-checkbox-marked-circle-outline:{ .lg .middle } __4. [Gebruik open standaarden](open-standaarden/index.md)__
-
-    ---
-
-    Bouw technologie die open standaarden gebruikt om ervoor te zorgen dat je technologie werkt en communiceert met andere technologie.
-
-- :material-cloud:{ .lg .middle } __5. [Hanteer een cloud-gedreven strategie](cloud/index.md)__
-
-    ---
-
-    Overweeg cloudoplossingen met aandacht voor Europese soevereiniteit en controle over data.
-
-- :material-shield-lock:{ .lg .middle } __6. [Maak veilige systemen](veiligheid/index.md)__
-
-    ---
-
-    Houd systemen en gegevens veilig met het juiste beveiligingsniveau.
-
-- :material-incognito:{ .lg .middle } __7. [Maak privacy integraal](privacy/index.md)__
-
-    ---
-
-    Zorg ervoor dat gebruikersrechten worden beschermd door privacy te integreren als een essentieel onderdeel van je systeem.
-
-- :material-share-variant:{ .lg .middle } __8. [Deel, hergebruik en werk samen](samenwerking/index.md)__
-
-    ---
-
-    Vermijd dubbel werk en onnodige kosten door samen te werken binnen de overheid en door technologie te delen en te hergebruiken.
-
-- :material-puzzle:{ .lg .middle } __9. [Integreer en pas technologie aan](integratie/index.md)__
-
-    ---
-
-    Je technologie moet werken met bestaande technologieën, processen en infrastructuur in je organisatie.
-
-- :material-database:{ .lg .middle } __10. [Maak beter gebruik van data](data/index.md)__
-
-    ---
-
-    Gebruik data effectiever door je technologie, infrastructuur en processen te verbeteren.
-
-- :material-robot:{ .lg .middle } __11. [Pas algoritmen verantwoord toe](algoritmen/index.md)__
-
-    ---
-
-    Zorg voor transparante, eerlijke en verantwoorde toepassing van algoritmen en AI-systemen.
-
-- :material-cart:{ .lg .middle } __12. [Definieer je inkoopstrategie](inkoop/index.md)__
-
-    ---
-
-    Je inkoopstrategie moet laten zien dat je commerciële en technologische aspecten hebt overwogen.
-
-- :material-leaf:{ .lg .middle } __13. [Maak je technologie duurzaam](duurzaamheid/index.md)__
-
-    ---
-
-    Verhoog duurzaamheid gedurende de hele levenscyclus van je technologie.
-
-- :material-check-all:{ .lg .middle } __14. [Voldoe aan de servicestandaard](servicestandaard/index.md)__
-
-    ---
-
-    Als je een dienst bouwt als onderdeel van je technologieproject of -programma, moet je ook voldoen aan de Servicestandaard.
-
-</div>
-
-Deze principes zijn gebaseerd op internationale best practices en aangepast aan de Nederlandse context. Ze helpen overheidsorganisaties bij het ontwikkelen van digitale systemen die gebruikersgericht, toegankelijk, veilig, duurzaam en kosteneffectief zijn.
