--- docs/principes/inkoop/index.md
@@ -1,120 +1,0 @@
----
-title: "12. Definieer je inkoopstrategie"
-summary: Je inkoopstrategie moet laten zien dat je commerciële en technologische aspecten hebt overwogen.
-relations:
-  - duurzaamheid
-  - cloud
-  - open-source
----
-
-# 12. Definieer je inkoopstrategie
-
-Een goed doordachte inkoopstrategie voor technologie zorgt ervoor dat je organisatie de juiste oplossingen verwerft tegen de beste waarde, met oog voor zowel huidige behoeften als toekomstige ontwikkelingen.
-
-## Waarom is het belangrijk?
-
-- **Waarde voor belastinggeld**: Zorg voor kosteneffectieve investeringen
-- **Vermijd vendor lock-in**: Behoud flexibiliteit en keuzevrijheid
-- **Risicobeheer**: Identificeer en beperk commerciële en technische risico's
-- **Innovatie stimuleren**: Creëer ruimte voor nieuwe oplossingen en aanbieders
-- **Duurzaamheid**: Overweeg milieu- en sociale impact van aankopen
-- **Strategische afstemming**: Zorg dat aankopen organisatiedoelen ondersteunen
-- **Wet- en regelgeving**: Voldoe aan aanbestedingsregels en andere regelgeving
-
-Een goed gedefinieerde inkoopstrategie voor technologie helpt je organisatie betere beslissingen te nemen, risico's te beheersen en waarde te creëren op zowel korte als lange termijn.
-
-## Hoe pas je het toe?
-
-<div class="direct-aan-de-slag">
-    <h3>Direct aan de slag</h3>
-
-    <div class="warning-banner" style="background-color: #fff3e0; padding: 0.5rem; border-left: 3px solid #ff9800; margin-bottom: 0.8rem;">
-        <strong>Work in Progress:</strong> De onderstaande functionaliteit is nog in ontwikkeling.
-    </div>
-
-    <div class="action-cards">
-        <div class="action-card">
-            <span class="wip-badge wip-badge-beschikbaar">beschikbaar</span>
-            <h4 >MVI-criteria</h4>
-            <p >Criteria voor duurzaam inkopen</p>
-            <a href="https://www.pianoo.nl/nl/themas/maatschappelijk-verantwoord-inkopen/productgroepen-en-mvi-criteria" class="action-button" target="_blank">Bekijken</a>
-        </div>
-        <div class="action-card">
-            <span class="wip-badge wip-badge-ontwikkeling">ontwikkeling</span>
-            <h4 >TCO Calculator</h4>
-            <p >Bereken Total Cost of Ownership</p>
-            <button disabled class="action-button">Berekenen</button>
-        </div>
-        <div class="action-card">
-            <span class="wip-badge wip-badge-beschikbaar">beschikbaar</span>
-            <h4 >Inkoopkompas</h4>
-            <p >Kennisplatform voor ICT-inkoop</p>
-            <a href="https://www.pianoo.nl/nl/sectoren/ict" class="action-button" target="_blank">Verkennen</a>
-        </div>
-        <div class="action-card">
-            <span class="wip-badge wip-badge-beschikbaar">beschikbaar</span>
-            <h4 >Algoritmekader</h4>
-            <p >Richtlijn voor inkoop van verantwoorde algoritmes</p>
-            <a href="https://minbzk.github.io/Algoritmekader/onderwerpen/publieke-inkoop/" class="action-button" target="_blank">Verkennen</a>
-        </div>
-    </div>
-</div>
-
-### Voorwaarde effectieve inkoopstrategie
-
- 1. **Behoefteanalyse**: Begrijp gebruikersbehoeften en organisatievereisten
- 2. **Marktverkenning**: Onderzoek beschikbare oplossingen en leveranciers
- 3. **Sourcingmodel**: Bepaal de juiste mix van intern/extern, cloud/on-premise
- 4. **Total cost of ownership**: Evalueer alle kosten over de volledige levenscyclus
- 5. **Risicoanalyse**: Identificeer en plan voor technische en commerciële risico's
- 6. **Contractstrategie**: Ontwikkel passende contractuele regelingen
- 7. **Exitstrategie**: Plan hoe je kunt overstappen naar andere oplossingen
-
-### Sleuteloverwegingen bij de ontwikkeling van je strategie
-
-- **Make vs. buy**: Wanneer zelf ontwikkelen en wanneer inkopen?
-- **Open source vs. commercieel**: Wat is de juiste balans?
-- **Grote vs. kleine leveranciers**: Hoe diversifieer je je leveranciersportfolio?
-- **Monolithisch vs. modulair**: Één grote of meerdere kleine oplossingen?
-- **Standaard vs. maatwerk**: Wanneer is aanpassing noodzakelijk?
-- **Korte vs. lange termijn**: Hoe balanceer je directe behoeften met toekomstbestendigheid?
-- **Centrale vs. decentrale inkoop**: Wat is het juiste niveau voor besluitvorming?
-
-Zie hiervoor ook [PIANOo's richtlijnen voor Maatschappelijk Verantwoord Inkopen](https://www.pianoo.nl/nl/themas/maatschappelijk-verantwoord-inkopen/productgroepen-en-mvi-criteria) voor productgroepen en MVI-criteria die je kunt toepassen bij je inkoopstrategie.
-
-### Best practices voor technologie-inkoop
-
-- **Vroege marktbetrokkenheid**: Ga in gesprek met leveranciers vóór formele aanbesteding
-- **Functionele specificaties**: Focus op wat je wilt bereiken, niet hoe
-- **Proof of concepts**: Test oplossingen voordat je je vastlegt
-- **Agile contracting**: Gebruik flexibele contractvormen voor softwareontwikkeling
-- **Regelmatige evaluatie**: Herzie contracten en relaties periodiek
-- **Samenwerking met leveranciers**: Bouw partnerschappen in plaats van strikt klant-leverancier relaties
-- **Kennisdeling**: Deel ervaringen binnen de overheid
-
-### Uitdagingen en oplossingen
-
-- **Complexe aanbestedingsregels**: Investeer in expertise en zoek innovatieve aanbestedingsvormen
-- **Snelle technologische verandering**: Bouw flexibiliteit in contracten
-- **Beperkte marktkennis**: Voer gedegen marktonderzoek uit
-- **Legacy-afhankelijkheden**: Ontwikkel migratiestrategieën
-- **Balanceren van meerdere belangen**: Stem af met alle belanghebbenden
-
-### Evaluatie van je inkoopstrategie
-
-- Is de strategie afgestemd op gebruikersbehoeften?
-- Houdt het rekening met de totale eigendomskosten?
-- Bevordert het concurrentie en innovatie?
-- Is er een duidelijke exit-strategie?
-- Is er aandacht voor duurzaamheid en sociale waarde?
-- Voldoet het aan alle wettelijke en beleidsmatige eisen?
-
-## Gerelateerde hulpmiddelen
-
-- [PIANOo's richtlijnen voor Maatschappelijk Verantwoord Inkopen](https://www.pianoo.nl/nl/themas/maatschappelijk-verantwoord-inkopen/productgroepen-en-mvi-criteria)
-
-## Gerelateerde principes
-
-- [3. Werk transparant en gebruik open source](../open-source/index.md)
-- [5. Hanteer een cloud-gedreven strategie](../cloud/index.md)
-- [13. Maak je technologie duurzaam](../duurzaamheid/index.md)
