--- docs/principes/open-source/index.md
@@ -1,120 +1,0 @@
----
-title: "3. Werk transparant en gebruik open source"
-summary: Publiceer je code en gebruik open source software om transparantie, flexibiliteit en verantwoording te verbeteren.
-relations:
-  - open-standaarden
-  - samenwerking
-  - veiligheid
----
-
-# 3. Werk transparant en gebruik open source
-
-Open source is een manier van werken waarbij de makers de broncode, documentatie en ontwerp openbaar beschikbaar stellen onder een opensourcelicentie. Hierdoor kan iedereen de software bekijken, kopiëren, aanpassen of delen. Mensen kunnen voorstellen doen voor verbeteringen of toevoegingen. Het delen van code en het gebruik van open source software kan leiden tot betere, veiligere en meer kosteneffectieve oplossingen.
-
-## Waarom is het belangrijk?
-
-- **Transparantie**: Iedereen kan zien hoe de software werkt en hoe besluiten worden genomen
-- **Kosteneffectiviteit**: Vermijd licentiekosten en afhankelijkheid van één leverancier
-- **Samenwerking**: Organisaties kunnen kennis delen en van elkaar leren
-- **Flexibiliteit**: Aanpassingen zijn mogelijk zonder afhankelijk te zijn van externe leveranciers
-- **Veiligheid**: Problemen kunnen sneller worden ontdekt en opgelost dankzij gemeenschappelijke inspectie
-- **Duurzaamheid**: Voorkomt vendor lock-in en maakt langetermijnbeheer mogelijk
-- **Innovatie**: Lost gemeenschappelijke problemen op met beschikbare technologie
-- **Tijd- en kostenbesparend**: Maakt gebruik van bestaande oplossingen in plaats van het wiel opnieuw uit te vinden
-
-Door open en transparant te werken en open source software te omarmen, draag je bij aan een meer transparante, flexibele en verantwoordelijke digitale overheid.
-
-## Hoe pas je het toe?
-
-<div class="direct-aan-de-slag">
-    <h3>Direct aan de slag</h3>
-
-    <div class="warning-banner" style="background-color: #fff3e0; padding: 0.5rem; border-left: 3px solid #ff9800; margin-bottom: 0.8rem;">
-        <strong>Work in Progress:</strong> De onderstaande functionaliteit is nog in ontwikkeling.
-    </div>
-
-    <div class="action-cards">
-        <div class="action-card">
-            <span class="wip-badge wip-badge-concept">concept</span>
-            <h4 >Code.gov.nl</h4>
-            <p >Publiceer uw code en vind overheidsprojecten</p>
-            <button disabled class="action-button">Publiceren</button>
-        </div>
-        <div class="action-card">
-            <span class="wip-badge wip-badge-beschikbaar">beschikbaar</span>
-            <h4 >Licentieadvies</h4>
-            <p >EUPL-1.2 voor overheidsprojecten</p>
-            <a href="https://ospo-nl.github.io/kennisbank/best-practices/open-source-software-licenties/" class="action-button" target="_blank">Bekijken</a>
-        </div>
-        <div class="action-card">
-            <span class="wip-badge wip-badge-ontwikkeling">ontwikkeling</span>
-            <h4 >OSPO Toolbox</h4>
-            <p >Tools voor Open Source Program Office</p>
-            <a href="https://ospo-nl.github.io/kennisbank/" class="action-button" target="_blank">Verkennen</a>
-        </div>
-        <div class="action-card">
-            <span class="wip-badge wip-badge-ontwikkeling">ontwikkeling</span>
-            <h4 >Opensourcewerken</h4>
-            <p >Community of practice omtrent open source werken</p>
-            <a href="https://developer.overheid.nl/communities/open-source-werken" class="action-button" target="_blank">Verkennen</a>
-        </div>
-    </div>
-</div>
-
-### Implementatiestappen
-
-1. **Evalueer bestaande oplossingen:** Onderzoek eerst of er al geschikte open source alternatieven bestaan.
-2. **Publiceer eigen code**: Maak de broncode van je software openbaar, tenzij er specifieke redenen zijn om dit niet te doen
-3. **Gebruik open standaarden**: Zorg ervoor dat je systemen kunnen communiceren met andere systemen (zie [Gebruik open standaarden](../open-standaarden/index.md))
-4. **Documenteer grondig**: Maak het voor anderen gemakkelijk om je code te begrijpen en eraan bij te dragen
-5. **Bouw een gemeenschap**: Moedig bijdragen en feedback aan op je projecten (zie [Deel, hergebruik en werk samen](../samenwerking/index.md))
-
-### Voordelen van het publiceren van je eigen code
-
-- **Betere documentatie**: Stimuleert duidelijkere documentatie
-- **Schonere code**: Leidt tot nettere, beter onderhoudbare code
-- **Dataveiligheid**: Zorgt voor duidelijkheid rond beschermde gegevens
-- **Verbetering**: Creëert mogelijkheden voor suggesties ter verbetering
-
-### Overwegingen bij het kiezen van open source technologie
-
-- **Voldoet het aan gebruikersbehoeften?** Beantwoordt de oplossing aan de specifieke eisen?
-- **Wat zijn de totale kosten?** Inclusief implementatie, migratie en ondersteuning
-- **Hoe volwassen is de code?** Is het een stabiel en actief onderhouden project?
-- **Welk ondersteuningsniveau bestaat er?** Is er een actieve gemeenschap of commerciële ondersteuning?
-- **Is het voldoende beveiligd?** Controleer veiligheidsgeschiedenis en updatefrequentie
-- **Is de licentie acceptabel?** Begrijp de licentievoorwaarden en -verplichtingen
-
-### Wanneer je eigen code open source maakt
-
-- Zorg ervoor dat er geen gevoelige informatie gepubliceerd wordt (zoals wachtwoorden en API-keys)
-- Zorg voor een duidelijke README met installatie en contributie-instructies
-- Kies een geschikte open source licentie (zie hieronder)
-- Onderhoud de documentatie actief
-- Wees responsief op issues en pull requests
-
-### Aanbevolen licentie voor overheidsprojecten
-
-Voor Nederlandse overheidsprojecten wordt de **European Union Public Licence (EUPL v1.2)** aanbevolen. Deze licentie:
-
-- Is specifiek ontwikkeld voor Europese overheidssoftware
-- Is juridisch gevalideerd in alle EU-talen
-- Heeft een "zwak wederkerig" karakter (wijzigingen in de originele code moeten onder dezelfde licentie worden gedeeld)
-- Is compatibel met andere populaire open source licenties
-- Voldoet aan de open source definitie
-
-Bij het publiceren van overheidscode is het raadzaam de EUPL-1.2 te gebruiken, tenzij er specifieke redenen zijn om een andere licentie te kiezen. Raadpleeg de [OSPO-NL kennisbank over OSS-licenties](https://ospo-nl.github.io/kennisbank/best-practices/open-source-software-licenties/) voor meer informatie over licentietypes en overwegingen.
-
-## Gerelateerde hulpmiddelen
-
-- [Opensourcewerken community](https://developer.overheid.nl/communities/open-source-werken) - Community van het Ministerie van Binnenlandse Zaken voor praktische kennis, een forum om ervaringen te delen en informatie over relevante regelgeving voor open source werken.
-
-- [OSPO-NL Kennisbank](https://ospo-nl.github.io/kennisbank/) - Verzameling van kennis over het opzetten van een Open Source Program Office (OSPO) binnen de Nederlandse overheid, met praktische handleidingen en richtlijnen voor het werken met open source.
-
-- [Instructie Rijksdienst bij aanschaf ICT-diensten of ICT-producten](https://wetten.overheid.nl/BWBR0024717/2008-11-23) - Officiële instructie die voorschrijft dat bij de aanschaf van ICT-diensten of -producten door de Rijksoverheid, open standaarden moeten worden geëist.
-
-## Gerelateerde principes
-
-- [4. Gebruik open standaarden](../open-standaarden/index.md)
-- [6. Maak veilige systemen](../veiligheid/index.md)
-- [8. Deel, hergebruik en werk samen](../samenwerking/index.md)
