--- docs/principes/servicestandaard/index.md
@@ -1,127 +1,0 @@
----
-title: "14. Voldoe aan de servicestandaard"
-summary: Als je een dienst bouwt als onderdeel van je technologieproject of -programma, moet je ook voldoen aan de Servicestandaard.
-relations:
-  - gebruikersbehoeften
-  - toegankelijkheid
-  - integratie
----
-
-# 14. Voldoe aan de servicestandaard
-
-De Servicestandaard definieert wat burgers en bedrijven mogen verwachten van overheidsdiensten. Het zorgt ervoor dat digitale diensten gebruiksvriendelijk, toegankelijk en betrouwbaar zijn, ongeacht welke overheidsorganisatie ze aanbiedt.
-
-## Waarom is het belangrijk?
-
-- **Consistente gebruikerservaring**: Zorgt voor een uniforme ervaring in alle overheidsdiensten
-- **Kwaliteitsborging**: Garandeert een minimumniveau van kwaliteit en bruikbaarheid
-- **Efficiëntie**: Voorkomt verspilling door slecht ontworpen diensten
-- **Inclusie**: Zorgt ervoor dat diensten voor iedereen toegankelijk zijn
-- **Vertrouwen**: Versterkt het vertrouwen van burgers in digitale overheidsdiensten
-- **Continu verbeteren**: Stimuleert voortdurende verbetering van diensten
-
-Door te voldoen aan de servicestandaard, leveren overheidsorganisaties diensten die efficiënt, gebruiksvriendelijk en inclusief zijn, wat leidt tot hogere gebruikerstevredenheid, lagere kosten en betere resultaten voor de samenleving.
-
-## Hoe pas je het toe?
-
-<div class="direct-aan-de-slag">
-    <h3>Direct aan de slag</h3>
-
-    <div class="warning-banner" style="background-color: #fff3e0; padding: 0.5rem; border-left: 3px solid #ff9800; margin-bottom: 0.8rem;">
-        <strong>Work in Progress:</strong> De onderstaande functionaliteit is nog in ontwikkeling.
-    </div>
-
-    <div class="action-cards">
-        <div class="action-card">
-            <span class="wip-badge wip-badge-concept">concept</span>
-            <h4 >Zelfassessment</h4>
-            <p >Beoordeel je dienst aan de standaard</p>
-            <button disabled class="action-button">Starten</button>
-        </div>
-        <div class="action-card">
-            <span class="wip-badge wip-badge-beschikbaar">beschikbaar</span>
-            <h4 >NL Design System</h4>
-            <p >Gebruiksvriendelijke componenten</p>
-            <a href="https://nldesignsystem.nl" class="action-button" target="_blank">Verkennen</a>
-        </div>
-        <div class="action-card">
-            <span class="wip-badge wip-badge-ontwikkeling">ontwikkeling</span>
-            <h4 >Service Pattern</h4>
-            <p >Gestandaardiseerde service patronen</p>
-            <button disabled class="action-button">Bekijken</button>
-        </div>
-    </div>
-</div>
-
-#### 1. Begrijp je gebruikers en hun behoeften
-
-- Onderzoek wie je gebruikers zijn en wat ze willen bereiken
-- Valideer aannames met gebruikersonderzoek
-- Blijf gebruikersfeedback verzamelen en verwerken
-
-#### 2. Los hele problemen op
-
-- Richt je op de volledige gebruikersreis, niet alleen het digitale deel
-- Werk over organisatiegrenzen heen
-- Vermijd geïsoleerde oplossingen die niet in de bredere context passen
-
-#### 3. Zorg voor een multidisciplinair team
-
-- Breng verschillende expertises samen (ontwerp, ontwikkeling, beleid, etc.)
-- Zorg voor directe betrokkenheid van alle benodigde disciplines
-- Stimuleer nauwe samenwerking in plaats van sequentiële overdracht
-
-#### 4. Creëer en itereer met gebruikers
-
-- Ontwikkel diensten incrementeel
-- Test vroeg en vaak met echte gebruikers
-- Verbeter continu op basis van feedback en gebruik
-
-#### 5. Zorg voor inclusie
-
-- Maak diensten die voor iedereen toegankelijk zijn
-- Houd rekening met verschillende vaardigheden, apparaten en omstandigheden
-- Bied alternatieven voor gebruikers die de digitale dienst niet kunnen gebruiken
-
-#### 6. Gebruik bewezen methoden en technologieën
-
-- Volg best practices en standaarden
-- Maak gebruik van bestaande, bewezen componenten
-- Vermijd onnodige complexiteit of experimentele technologieën
-
-#### 7. Opereer betrouwbaar en veilig
-
-- Bouw robuuste, betrouwbare diensten
-- Implementeer passende beveiligingsmaatregelen
-- Zorg voor transparantie over datagebruik en privacy
-
-### Het beoordelingsproces
-
-- **Zelfevaluatie**: Beoordeel je eigen dienst aan de hand van de standaard
-- **Onafhankelijke toetsing**: Laat je dienst beoordelen door experts
-- **Continu verbeteren**: Gebruik beoordelingen om diensten te verbeteren
-- **Publicatie van resultaten**: Transparantie over hoe diensten presteren
-
-### Implementatie in de praktijk
-
-- Start met de servicestandaard vanaf het begin van het project
-- Documenteer hoe je aan de standaard voldoet
-- Gebruik assessments als leermomenten, niet alleen als controlepunten
-- Deel ervaringen en lessen binnen en buiten je organisatie
-- Bouw een cultuur van gebruikersgerichtheid en continue verbetering
-
-### Uitdagingen en oplossingen
-
-- **Legacy systemen**: Ontwikkel een strategie voor geleidelijke verbetering
-- **Organisatorische grenzen**: Bevorder samenwerking tussen afdelingen
-- **Beperkte middelen**: Focus op de meest kritieke aspecten eerst
-- **Veranderingsmanagement**: Investeer in cultuurverandering en training
-- **Complexe regelgeving**: Vertaal complexiteit naar gebruiksvriendelijke ervaringen
-
-## Gerelateerde standaarden
-
-## Gerelateerde principes
-
-- [1. Stel gebruikersbehoeften vast](../gebruikersbehoeften/index.md)
-- [2. Zorg voor toegankelijkheid en inclusie](../toegankelijkheid/index.md)
-- [9. Integreer en pas technologie aan](../integratie/index.md)
