--- docs/principes/toegankelijkheid/index.md
@@ -1,110 +1,0 @@
----
-title: "2. Zorg voor toegankelijkheid en inclusie"
-summary: Zorg ervoor dat je technologie, infrastructuur en systemen toegankelijk en inclusief zijn voor alle gebruikers.
-relations:
-  - gebruikersbehoeften
-  - servicestandaard
-  - open-standaarden
----
-
-# 2. Zorg voor toegankelijkheid en inclusie
-
-Toegankelijkheid gaat over het maken van diensten en systemen die door iedereen kunnen worden gebruikt, inclusief mensen met handicaps of specifieke behoeften. Dit principe zorgt ervoor dat alle burgers gelijke toegang hebben tot overheidsdiensten.
-
-## Waarom is het belangrijk?
-
-- **Wettelijke verplichting**: Overheidsorganisaties zijn wettelijk verplicht om hun digitale diensten toegankelijk te maken
-- **Beter voor iedereen**: Toegankelijke diensten zijn meestal eenvoudiger en gebruiksvriendelijker voor iedereen
-- **Bereik vergroten**: Je bereikt een breder publiek en zorgt dat niemand wordt uitgesloten
-- **Kosteneffectiviteit**: Toegankelijkheid vanaf het begin inbouwen is goedkoper dan achteraf aanpassen
-- **Ondersteuning van personeel**: Stelt alle medewerkers in staat om de benodigde infrastructuur te gebruiken
-- **Vermijd juridische uitdagingen**: Voorkom potentiële juridische problemen door niet-toegankelijke technologie
-
-Door toegankelijkheid vanaf het begin in te bouwen in je digitale systemen, zorg je ervoor dat je diensten voor iedereen bruikbaar zijn, ongeacht hun capaciteiten of beperkingen.
-
-## Hoe pas je het toe?
-
-<div class="direct-aan-de-slag">
-    <h3>Direct aan de slag</h3>
-
-    <div class="warning-banner" style="background-color: #fff3e0; padding: 0.5rem; border-left: 3px solid #ff9800; margin-bottom: 0.8rem;">
-        <strong>Work in Progress:</strong> De onderstaande functionaliteit is nog in ontwikkeling.
-    </div>
-
-    <div class="action-cards">
-        <div class="action-card">
-            <span class="wip-badge wip-badge-beschikbaar">beschikbaar</span>
-            <h4 >Toegankelijkheidsscan</h4>
-            <p >Test uw webpagina's op toegankelijkheid</p>
-            <a href="https://www.w3.org/WAI/test-evaluate/tools/list/" class="action-button" target="_blank">Scannen</a>
-        </div>
-        <div class="action-card">
-            <span class="wip-badge wip-badge-beschikbaar">beschikbaar</span>
-            <h4 >Deque Axe</h4>
-            <p >Toegankelijkheidstester voor webapplicaties</p>
-            <a href="https://www.deque.com/axe/" class="action-button" target="_blank">Gebruiken</a>
-        </div>
-        <div class="action-card">
-            <span class="wip-badge wip-badge-beschikbaar">beschikbaar</span>
-            <h4 >WCAG-EM Reporter</h4>
-            <p >Tool voor het maken van toegankelijkheidsrapporten</p>
-            <a href="https://gitlab.com/digilab.overheid.nl/ecosystem/wcag-em-reporter" class="action-button" target="_blank">Gebruiken</a>
-        </div>
-        <div class="action-card">
-            <span class="wip-badge wip-badge-ontwikkeling">ontwikkeling</span>
-            <h4 >Inclusief Ontwerpen</h4>
-            <p >Toolkit voor inclusief ontwerp</p>
-            <a href="https://toolkitinclusie.gebruikercentraal.nl/" class="action-button" target="_blank">Bekijken</a>
-        </div>
-        <div class="action-card">
-            <span class="wip-badge wip-badge-beschikbaar">beschikbaar</span>
-            <h4 >NL Design System</h4>
-            <p >Toegankelijke componenten</p>
-            <a href="https://nldesignsystem.nl" class="action-button" target="_blank">Gebruiken</a>
-        </div>
-    </div>
-</div>
-
-1. **Houd rekening met toegankelijkheid in**:
-    - Kantoor- en serverruimtes
-    - Hardwarecompatibiliteit
-    - Softwareselectie
-    - Gebruikersinterfaceontwerp
-
-2. **Aanbevolen praktijken**:
-    - Volg de webrichtlijnen (WCAG 2.2 AA-standaard minimaal)
-    - Test met hulptechnologieën
-    - Gebruik eenvoudige taal
-    - Zorg voor alternatieve formaten
-
-### Inkopen van toegankelijke technologie
-
-- Gebruik de EN 301 549 toegankelijkheidsstandaard in aanbestedingen
-- Zoek naar leveranciers met bewezen toegankelijkheidservaring
-- Vraag om bewijs van toegankelijkheidstesten
-- Zorg dat toekomstige updates toegankelijkheid behouden
-
-### Belangrijke overwegingen
-
-- Zorg dat systemen werken met verschillende invoermethoden (toetsenbord, muis, aanraakscherm, spraak)
-- Let op kleurcontrast voor mensen met visuele beperkingen
-- Zorg voor ondertiteling en transcripties bij audio- en video-inhoud
-- Bouw formulieren die gemakkelijk te gebruiken zijn met hulpapparatuur
-- Test je systemen op verschillende apparaten en schermformaten
-
-## Gerelateerde standaarden
-
-- [Web Content Accessibility Guidelines (WCAG 2.2)](https://www.w3.org/TR/WCAG22/)
-- [EN 301 549](https://www.etsi.org/deliver/etsi_en/301500_301599/301549/03.02.01_60/en_301549v030201p.pdf) (Europese norm voor hardware/niet-web software)
-- [Digitoegankelijkheid](https://www.digitoegankelijk.nl/toegankelijkheid/en-301-549-en-wcag) (voorheen 'Webrichtlijnen')
-
-## Gerelateerde hulpmiddelen
-
-- [Handreiking Mobiele app Ontwikkeling en Beheer versie 4.0](https://www.noraonline.nl/images/noraonline/2/26/Handreiking_Mobiele_App_4.0.pdf) - Hoofdstuk 8.3 geeft concrete best practices voor verschillende beperkingen voor de ontwikkeling van Mobiele Apps
-- [Kennisbank Digitoegankelijk](https://kennisbank.digitoegankelijk.nl) - Praktische tips over digitale toegankelijkheid
-
-## Gerelateerde principes
-
-- [1. Stel gebruikersbehoeften vast](../gebruikersbehoeften/index.md)
-- [4. Gebruik open standaarden](../open-standaarden/index.md)
-- [14. Voldoe aan de servicestandaard](../servicestandaard/index.md)
