--- docs/principes/veiligheid/index.md
@@ -1,136 +1,0 @@
----
-title: "6. Maak veilige systemen"
-summary: Houd systemen en gegevens veilig met het juiste beveiligingsniveau.
-relations:
-  - privacy
-  - data
-  - algoritmen
-  - cloud
----
-
-# 6. Maak veilige systemen
-
-Beveiliging is een fundamenteel aspect van digitale systemen. Het gaat niet alleen om het beschermen tegen aanvallen, maar ook om het waarborgen van de beschikbaarheid, integriteit en vertrouwelijkheid van gegevens en diensten.
-
-## Waarom is het belangrijk?
-
-- **Bescherming van gevoelige gegevens**: Overheidsorganisaties verwerken vaak persoonlijke en gevoelige informatie
-- **Behoud van vertrouwen**: Burgers moeten erop kunnen vertrouwen dat hun gegevens veilig zijn
-- **Wettelijke verplichting**: Naleving van relevante wetgeving en standaarden
-- **Continuïteit van dienstverlening**: Voorkom verstoring van essentiële overheidsdiensten
-- **Bescherming tegen bedreigingen**: Verdediging tegen een groeiend aantal cyberdreigingen
-- **Kostenbeheersing**: Voorkom dure herstelkosten en boetes na beveiligingsincidenten
-- **Reputatiebescherming**: Behoud het vertrouwen van burgers en partners in de overheid
-- **Concurrentievoordeel**: Sterke beveiliging maakt innovatie en samenwerking mogelijk
-
-Door beveiliging serieus te nemen en het te integreren in alle aspecten van technologieontwikkeling en -gebruik, bescherm je niet alleen gegevens maar ook het vertrouwen van burgers in de digitale overheid.
-
-## Hoe pas je het toe?
-
-<div class="direct-aan-de-slag">
-    <h3>Direct aan de slag</h3>
-
-    <div class="warning-banner" style="background-color: #fff3e0; padding: 0.5rem; border-left: 3px solid #ff9800; margin-bottom: 0.8rem;">
-        <strong>Work in Progress:</strong> De onderstaande functionaliteit is nog in ontwikkeling.
-    </div>
-
-    <div class="action-cards">
-        <div class="action-card">
-            <span class="wip-badge wip-badge-beschikbaar">beschikbaar</span>
-            <h4 >BIO Wizard</h4>
-            <p >Bepaal verplichte beveiligingsmaatregelen</p>
-            <a href="https://bio-overheid.nl/ico-wizard" class="action-button" target="_blank">Starten</a>
-        </div>
-        <div class="action-card">
-            <span class="wip-badge wip-badge-beschikbaar">beschikbaar</span>
-            <h4 >OpenKAT</h4>
-            <p >Open source security tooling</p>
-            <a href="https://github.com/minvws/nl-kat-coordination" class="action-button" target="_blank">Gebruiken</a>
-        </div>
-        <div class="action-card">
-            <span class="wip-badge wip-badge-beschikbaar">beschikbaar</span>
-            <h4 >Internet.nl</h4>
-            <p >Test veiligheid van je website of e-mail</p>
-            <a href="https://internet.nl/" class="action-button" target="_blank">Testen</a>
-        </div>
-    </div>
-</div>
-
-1. **Overweeg beveiliging vanaf het begin**:
-    - Beoordeel beveiligingsrisico's voordat je aan een technologieprogramma begint
-    - Identificeer potentiële kwetsbaarheden in systemen en gegevens
-    - Bepaal welke beveiligingsexpertise en middelen nodig zijn
-
-2. **Implementeer proportionele beveiligingsmaatregelen**:
-    - Bescherm informatietechnologie en digitale diensten
-    - Zorg voor passende toegang tot gebruikersgegevens
-    - Evalueer en update voortdurend beveiligingsprotocollen
-
-3. **Focus op meerdere beveiligingsdimensies**:
-    - Netwerk- en infrastructuurbeveiliging
-    - Gegevensbeveiliging
-    - Dienstbeveiliging
-    - Cloudbeveiliging
-
-### Kritieke beveiligingsstrategieën
-
-- **Identificeer, bescherm, detecteer, reageer en herstel**: Volg deze cyclus voor beveiligingsincidenten
-- **Gebruik beveiligingscontroles zoals**:
-  - Gegevensversleuteling
-  - Single sign-on
-  - Tweefactorauthenticatie
-  - Specifieke toegangscontrole
-  - Gebruiksmonitoring
-  - Tijdige systeempatching
-
-### Implementatie van beveiligingsmaatregelen
-
-- **Risicoanalyse**: Identificeer bedreigingen en kwetsbaarheden
-- **Authenticatie en autorisatie**: Implementeer sterke toegangscontroles
-- **Encryptie**: Bescherm gegevens tijdens opslag en transport
-- **Veilige ontwikkelpraktijken**: Volg secure coding standaarden
-- **Penetratietesten**: Test regelmatig de beveiliging van systemen
-- **Incident response**: Ontwikkel en test plannen voor het reageren op incidenten
-- **Leveranciersbeheer**: Evalueer en controleer de beveiliging van leveranciers
-
-### Doorlopend beveiligingsbeheer
-
-- Stel plannen voor continue verbetering op
-- Beoordeel regelmatig beveiligingscontroles
-- Monitor potentiële risico's
-- Pas aan aan veranderende technologische landschappen
-
-### Aanbevolen acties
-
-- Voer uitgebreide risicobeoordelingen uit
-- Wijs duidelijke beveiligingsrollen en -verantwoordelijkheden toe
-- Documenteer beveiligingsprocessen
-- Creëer gebruikerstoegang training en controleplannen
-- Integreer beveiligingsoverwegingen in algemene programmaplanningen
-
-## Gerelateerde standaarden
-
-- [ISO 27001](https://www.forumstandaardisatie.nl/open-standaarden/nen-isoiec-27001) en [ISO 27002](https://www.forumstandaardisatie.nl/open-standaarden/nen-isoiec-27002)
-- [Baseline Informatiebeveiliging Overheid (BIO)](https://bio-overheid.nl/ico-wizard) - Gebruik de BIO ICO-wizard om te bepalen welke BIO-maatregelen in jouw situatie van toepassing zijn
-- [NIST Cybersecurity Framework](https://nvlpubs.nist.gov/nistpubs/CSWP/NIST.CSWP.29.pdf)
-- [OWASP Top 10 voor webapplicatiebeveiliging](https://owasp.org/www-project-top-ten/)
-- [Privacy by Design principes](https://www.rijksoverheid.nl/documenten/rapporten/2023/06/01/handleiding-privacy-by-design-versie-1-0)
-
-## Gerelateerde hulpmiddelen
-
-- [Nationaal Cyber Security Centrum (NCSC)](https://www.ncsc.nl/) - Centrale kennisautoriteit voor cybersecurity in Nederland, met actuele dreigingsinformatie, advisering en ondersteuning
-  - [NCSC ICT-beveiligingsrichtlijnen voor webapplicaties](https://www.ncsc.nl/documenten/publicaties/2019/mei/01/ict-beveiligingsrichtlijnen-voor-webapplicaties)
-  - [NCSC ICT-beveiligingsrichtlijnen voor mobiele apparaten](https://www.ncsc.nl/documenten/publicaties/2019/mei/01/beveiligingsrichtlijnen-voor-mobiele-apparaten)
-- [Rijksinspectie Digitale Infrastructuur (RDI)](https://www.rdi.nl/) - Toezichthouder op het gebied van digitale infrastructuur en telecom
-- [Nationaal innovatie centrum privacy-enhancing technologies (NICPET)](https://nicpet.pleio.nl/) - Innovatiecentrum gericht op privacybeschermende technologieën
-- [OpenKAT](https://github.com/minvws/nl-kat-coordination) - Open source security kennisbank en tooling voor automatische testen
-- [Internet.nl](https://internet.nl/) - Test of je website en e-mail moderne internetstandaarden gebruiken die helpen om te beschermen tegen phishing en afluisteren
-- [Handreiking Mobiele app Ontwikkeling en Beheer versie 4.0](https://www.noraonline.nl/images/noraonline/2/26/Handreiking_Mobiele_App_4.0.pdf) - Hoofdstuk 10 geeft concrete beveiligingsmaatregelen voor de ontwikkeling van mobiele apps
-- [Handreiking Cybersecurity voor Bestuurders en Bedrijfseigenaren](https://www.cybersecurityraad.nl/documenten/2025/06/04/handreiking-cybersecurity-voor-bestuurders-en-bedrijfseigenaren) <!--RM: ik weet niet zeker of deze bron nuttig is-->
-
-## Gerelateerde principes
-
-- [7. Maak privacy integraal](../privacy/index.md)
-- [10. Maak beter gebruik van data](../data/index.md)
-- [11. Pas algoritmen verantwoord toe](../algoritmen/index.md)
-- [5. Hanteer een cloud-gedreven strategie](../cloud/index.md)
