+++ docs/richtlijnen/data/index.md
@@ -1,0 +1,112 @@
+---
+title: "10. Maak beter gebruik van data"
+summary: Gebruik data effectiever door je technologie, infrastructuur en processen te verbeteren.
+relations:
+  - algoritmen
+  - privacy
+  - veiligheid
+  - open-standaarden
+---
+
+# 10. Maak beter gebruik van data
+
+Data is een waardevolle hulpbron die, wanneer effectief gebruikt en beheert, kan leiden tot betere besluitvorming, efficiëntere diensten en innovatieve oplossingen voor complexe problemen.
+
+## Waarom is het belangrijk?
+
+- **Geïnformeerde besluitvorming**: Onderbouw beleid en strategieën met feitelijke gegevens
+- **Verbeterde dienstverlening**: Pas diensten aan op basis van gebruikspatronen en behoeften
+- **Efficiëntere processen**: Identificeer knelpunten en optimaliseer werkstromen
+- **Innovatie**: Ontdek nieuwe inzichten en mogelijkheden
+- **Transparantie**: Verhoog verantwoording door gegevens te delen
+- **Voorspellend vermogen**: Anticipeer op toekomstige trends en behoeften
+
+Door data effectiever te gebruiken, kunnen overheidsorganisaties betere diensten leveren, efficiënter werken en de complexe uitdagingen van vandaag en morgen beter aanpakken.
+
+## Hoe pas je het toe?
+
+<div class="direct-aan-de-slag">
+    <h3>Direct aan de slag</h3>
+
+    <div class="warning-banner" style="background-color: #fff3e0; padding: 0.5rem; border-left: 3px solid #ff9800; margin-bottom: 0.8rem;">
+        <strong>Work in Progress:</strong> De onderstaande functionaliteit is nog in ontwikkeling.
+    </div>
+
+    <div class="action-cards">
+        <div class="action-card">
+            <span class="wip-badge wip-badge-beschikbaar">beschikbaar</span>
+            <h4 >Open Data</h4>
+            <p >Het open dataportaal van de overheid</p>
+            <a href="https://data.overheid.nl/" class="action-button" target="_blank">Verkennen</a>
+        </div>
+        <div class="action-card">
+            <span class="wip-badge wip-badge-ontwikkeling">ontwikkeling</span>
+            <h4 >Datacheck</h4>
+            <p >Tool voor datakwaliteitscontrole</p>
+            <button disabled class="action-button">Starten</button>
+        </div>
+        <div class="action-card">
+            <span class="wip-badge wip-badge-concept">concept</span>
+            <h4 >Synthetische Data</h4>
+            <p >Genereer veilige testdata</p>
+            <a href="https://digilab.overheid.nl/projecten/synthetische-datageneratie/" class="action-button" target="_blank">Leren</a>
+        </div>
+    </div>
+</div>
+
+### Implementatiestappen
+
+- **Data-inventarisatie**: Breng in kaart welke gegevens beschikbaar zijn
+- **Data management strategie**: Ontwikkel een alomvattende aanpak
+- **Datakwaliteitscontroles**: Implementeer processen voor kwaliteitsborging
+- **Metadata management**: Documenteer de betekenis en context van gegevens
+- **Data-integratie**: Verbind gegevensbronnen voor een completer beeld
+- **Analytics capabiliteit**: Bouw capaciteit voor gegevensanalyse
+- **Open data initiatieven**: Deel relevante gegevens met het publiek
+
+### Technologische enablers
+
+- **Data lakes en warehouses**: Centrale opslag voor diverse gegevenstypen
+- **Business intelligence tools**: Software voor analyse en visualisatie
+- **API's voor gegevenstoegang**: Gestructureerde toegang tot gegevens
+- **ETL-processen**: Extractie, transformatie en laden van gegevens
+- **Machine learning**: Geavanceerde analyse voor complexe patronen
+- **Real-time verwerking**: Systemen voor directe gegevensverwerking
+- **Datavisualisatie**: Tools om gegevens begrijpelijk te presenteren
+
+### Uitdagingen en oplossingen
+
+- **Datasilo's**: Bevorder organisatiebrede samenwerking
+- **Privacybescherming**: Implementeer privacy by design
+- **Legacy systemen**: Ontwikkel strategieën voor data-extractie
+- **Datavolume en -complexiteit**: Gebruik schaalbare oplossingen
+- **Databeheer**: Implementeer levenscyclus- en retentiebeleid
+- **Vaardigheidstekorten**: Investeer in training en expertise
+
+### Succesfactoren
+
+- Duidelijke koppeling tussen datastrategieën en organisatiedoelen
+- Leiderschap en cultuur die databewuste besluitvorming stimuleren
+- Adequate investeringen in datamanagement en -infrastructuur
+- Balans tussen innovatie en compliance
+- Voortdurende evaluatie en verbetering van datapraktijken
+- Samenwerking tussen technische teams en domeinexperts
+
+## Gerelateerde hulpmiddelen
+
+- **[Uit Betrouwbare Bron](https://digilab.overheid.nl/projecten/)**: Initiatief gericht op datakwaliteit in federatieve datasystemen, met nadruk op het richtlijn 'data bij de bron'.
+
+- **[Logboek Dataverwerkingen](https://digilab.overheid.nl/projecten/)**: Creëert transparantie in overheidsdata door bij te houden wie data wanneer en waarom verandert.
+
+- **[Stelsel van Basisregistraties](https://www.digitaleoverheid.nl/overzicht-van-alle-onderwerpen/stelsel-van-basisregistraties/)**: Samenhangend stelsel van hoogwaardige registraties met essentiële overheidsgegevens.
+
+- **[RSGB/RGBZ](https://www.gemmaonline.nl/wiki/Hoofdpagina)**: Referentiemodellen voor gemeentelijke basis- en zaakgegevens, die zorgen voor consistente definitie en uitwisseling van data.
+
+- **[data.overheid.nl](https://data.overheid.nl/)**: Het open data portaal van de Nederlandse overheid, dat toegang biedt tot datasets van overheidsorganisaties.
+
+## Gerelateerde richtlijnen
+
+- [4. Gebruik open standaarden](../open-standaarden/index.md)
+- [6. Maak veilige systemen](../veiligheid/index.md)
+- [7. Maak privacy integraal](../privacy/index.md)
+- [11. Pas algoritmen verantwoord toe](../algoritmen/index.md)
