+++ docs/richtlijnen/duurzaamheid/index.md
@@ -1,0 +1,131 @@
+---
+title: "13. Maak je technologie duurzaam"
+summary: Verhoog duurzaamheid gedurende de hele levenscyclus van je technologie.
+relations:
+  - cloud
+  - samenwerking
+  - inkoop
+---
+
+# 13. Maak je technologie duurzaam
+
+Duurzame technologie minimaliseert negatieve milieueffecten, optimaliseert energieverbruik en draagt bij aan bredere duurzaamheidsdoelstellingen. Door duurzaamheid te integreren in de volledige levenscyclus van technologie, dragen overheidsorganisaties bij aan een groenere toekomst.
+
+## Waarom is het belangrijk?
+
+- **Klimaatverandering**: Vermindering van broeikasgasemissies
+- **Energieverbruik**: Besparing op kosten en vermindering van milieubelasting
+- **Afvalbeperking**: Vermindering van elektronisch afval
+- **Grondstoffenschaarste**: Verantwoord gebruik van zeldzame materialen
+- **Voorbeeldfunctie**: De overheid als aanjager van duurzame praktijken
+- **Regelgeving**: Anticiperen op strengere milieunormen
+- **Kostenbesparing**: Efficiënter gebruik van middelen op lange termijn
+
+Door technologische duurzaamheid te prioriteren, kunnen overheidsorganisaties hun ecologische voetafdruk verkleinen, kosten besparen en bijdragen aan bredere duurzaamheidsdoelstellingen.
+
+## Hoe pas je het toe?
+
+<div class="direct-aan-de-slag">
+    <h3>Direct aan de slag</h3>
+
+    <div class="warning-banner" style="background-color: #fff3e0; padding: 0.5rem; border-left: 3px solid #ff9800; margin-bottom: 0.8rem;">
+        <strong>Work in Progress:</strong> De onderstaande functionaliteit is nog in ontwikkeling.
+    </div>
+
+    <div class="action-cards">
+        <div class="action-card">
+            <span class="wip-badge wip-badge-beschikbaar">beschikbaar</span>
+            <h4 >ICT MVI-criteria</h4>
+            <p >Duurzaamheidscriteria voor ICT-inkoop</p>
+            <a href="https://www.pianoo.nl/nl/themas/maatschappelijk-verantwoord-inkopen/productgroepen-en-mvi-criteria" class="action-button" target="_blank">Toepassen</a>
+        </div>
+        <div class="action-card">
+            <span class="wip-badge wip-badge-concept">concept</span>
+            <h4 >CO2-calculator</h4>
+            <p >Bereken de CO2-impact van uw IT</p>
+            <a href="https://www.metenvanduurzaamheid.nl/overzicht-tools/monitoring-als-tool-voor-bedrijven-en-organisaties/co2-prestatieladder-en-co2-calculator" class="action-button" target="_blank">Verkennen</a>
+        </div>
+        <div class="action-card">
+            <span class="wip-badge wip-badge-ontwikkeling">ontwikkeling</span>
+            <h4 >Green Coding</h4>
+            <p >Best practices voor duurzame software</p>
+            <button disabled class="action-button">Verkennen</button>
+        </div>
+    </div>
+</div>
+
+#### 1. Planning en ontwerp
+
+- Neem duurzaamheidseisen op in programma's van eisen
+- Overweeg de milieu-impact bij het vergelijken van alternatieven
+- Ontwerp voor energiezuinigheid en lange levensduur
+- Denk na over de end-of-life fase vanaf het begin
+
+#### 2. Inkoop en productie
+
+- Neem duurzaamheidscriteria op in aanbestedingen
+- Beoordeel leveranciers op hun milieubeleid en prestaties
+- Kies voor producten met milieukeurmerken
+- Selecteer leveranciers met terugname- en recyclingprogramma's
+- Raadpleeg [PIANOo's MVI-criteria voor ICT](https://www.pianoo.nl/nl/themas/maatschappelijk-verantwoord-inkopen/productgroepen-en-mvi-criteria) voor specifieke richtlijnen bij aanbestedingen
+
+#### 3. Gebruik en beheer
+
+- Optimaliseer energiegebruik van hardware en software
+- Implementeer energiebesparende instellingen
+- Monitor en rapporteer over energieverbruik
+- Train gebruikers in duurzaam gebruik van technologie
+
+#### 4. Einde levensduur
+
+- Maximaliseer hergebruik van apparatuur
+- Zorg voor verantwoorde recycling
+- Wis gegevens veilig voordat apparatuur wordt afgedankt
+- Documenteer de milieuvoordelen van vervanging vs. verlenging
+
+### Strategieën voor duurzame technologie
+
+- **Cloudoptimalisatie**: Kies voor energie-efficiënte cloudproviders
+- **Green coding**: Ontwikkel software die efficiënter met resources omgaat
+- **Hardware levensduurverlenging**: Upgrade in plaats van vervang waar mogelijk
+- **Virtualisatie**: Consolideer servers en infrastructuur
+- **Circulair inkopen**: Koop refurbished of herbruikbare apparatuur
+- **Groene datacenters**: Kies voor faciliteiten die duurzame energie gebruiken
+- **Papierloos werken**: Digitaliseer processen om papierverbruik te verminderen
+- **Slimme gebouwen**: Gebruik technologie voor energiebesparing in gebouwen
+
+### Meten en verbeteren van duurzaamheid
+
+- **Baseline vaststellen**: Meet huidige energie- en materiaalverbruik
+- **KPI's definiëren**: Stel doelen voor verbetering
+- **Levensduurfactoren**: Beoordeel total cost of ownership inclusief milieukosten
+- **Rapportage**: Publiceer duurzaamheidsprestaties
+- **Certificeringen**: Streef naar erkende milieustandaarden
+- **Continu verbeteren**: Evalueer en pas aan op basis van resultaten
+
+### Uitdagingen en oplossingen
+
+- **Initiële kosten**: Focus op totale kosten over de levensduur
+- **Gebrek aan kennis**: Investeer in training en bewustwording
+- **Legacy systemen**: Ontwikkel geleidelijke verduurzamingsstrategieën
+- **Complexe leveranciersketens**: Werk samen met kernleveranciers
+- **Balanceren van prestaties**: Zoek optimale balans tussen functionaliteit en duurzaamheid
+
+### Best practices
+
+- Integreer duurzaamheid in IT-governance
+- Ontwikkel een duurzaamheidsroadmap met mijlpalen
+- Betrek gebruikers bij duurzaamheidsinitiatieven
+- Deel successen en lessen binnen en buiten de organisatie
+- Werk samen met andere overheidsorganisaties aan gemeenschappelijke doelen
+
+## Gerelateerde hulpmiddelen
+
+- [Duurzame en sociale doelen van de overheid](https://www.denkdoeduurzaam.nl/doelen)
+- [PIANOo's MVI-criteria voor ICT](https://www.pianoo.nl/nl/themas/maatschappelijk-verantwoord-inkopen/productgroepen-en-mvi-criteria)
+
+## Gerelateerde richtlijnen
+
+- [5. Hanteer een cloud-gedreven strategie](../cloud/index.md)
+- [8. Deel, hergebruik en werk samen](../samenwerking/index.md)
+- [12. Definieer je inkoopstrategie](../inkoop/index.md)
