+++ docs/richtlijnen/gebruikersbehoeften/index.md
@@ -1,0 +1,86 @@
+---
+title: "1. Stel gebruikersbehoeften vast"
+summary: Begrijp je gebruikers en hun behoeften. Ontwikkel kennis over je gebruikers en wat dat betekent voor je technologieproject.
+relations:
+  - toegankelijkheid
+  - privacy
+  - servicestandaard
+---
+
+# 1. Stel gebruikersbehoeften vast
+
+Om digitale systemen te ontwikkelen die effectief zijn, moet je je gebruikers en hun behoeften begrijpen. Dit geldt vooral wanneer je door een uitgavecontroleproces gaat.
+
+## Waarom is het belangrijk?
+
+- **Voorkom kostbare fouten:** Bouw vanaf het begin de juiste oplossing in plaats van achteraf aanpassen
+- **Verhoog adoptie:** Gebruikers omarmen systemen die hun werkelijke problemen oplossen
+- **Reduceer risico's:** Identificeer potentiële problemen voordat technologie wordt geïmplementeerd
+- **Bespaar tijd en geld:** Voorkom ontwikkeling van ongewenste features
+- **Verbeter dienstverlening:** Systemen die aansluiten bij gebruikersbehoeften werken effectiever
+- **Ondersteun interne alignment:** Creëer overeenstemming over technologiedoelstellingen binnen teams
+
+Door een gebruikersgerichte benadering toe te passen van technologieontwikkeling en -implementatie bij overheidsdiensten verhoog je de acceptatie en effectiviteit van digitale overheidsdiensten terwijl je het vertrouwen tussen burger en overheid versterkt.
+
+## Hoe pas je het toe?
+
+<div class="direct-aan-de-slag">
+    <h3>Direct aan de slag</h3>
+
+    <div class="warning-banner" style="background-color: #fff3e0; padding: 0.5rem; border-left: 3px solid #ff9800; margin-bottom: 0.8rem;">
+        <strong>Work in Progress:</strong> De onderstaande functionaliteit is nog in ontwikkeling.
+    </div>
+
+    <div class="action-cards">
+        <div class="action-card">
+            <span class="wip-badge wip-badge-concept">concept</span>
+            <h4 >Gebruikers Panel</h4>
+            <p >Toegang tot gebruikerspanel van burgers voor onderzoek</p>
+            <button disabled class="action-button">Aanvragen</button>
+        </div>
+        <div class="action-card">
+            <span class="wip-badge wip-badge-ontwikkeling">ontwikkeling</span>
+            <h4 >Persona Database</h4>
+            <p >Gevalideerde overheidsgebruikerspersona's</p>
+            <button disabled class="action-button">Verkennen</button>
+        </div>
+        <div class="action-card">
+            <span class="wip-badge wip-badge-beschikbaar">beschikbaar</span>
+            <h4 >Onderzoeksrichtlijnen</h4>
+            <p >Toolkit voor gebruikersonderzoek</p>
+            <a href="https://toolkittaal.gebruikercentraal.nl/richtlijnen/" class="action-button" target="_blank">Bekijken</a>
+        </div>
+    </div>
+</div>
+
+### Onderzoek uitvoeren
+
+ Gebruikersonderzoek kan helpen bij het identificeren van:
+
+- Risico's bij het introduceren of wijzigen van technologie
+- Vaardigheden die nodig zijn om technologie te leveren, gebruiken en beheren
+- Technologieën die nodig zijn voor serviceteams
+- Commerciële en operationele behoeften (bijvoorbeeld het buiten gebruik stellen van verouderde systemen)
+
+### Aanvullende voordelen van gebruikersonderzoek
+
+- Zorg ervoor dat diensten zoals kantoorsoftware en projectmanagementtools voldoen aan gebruikersbehoeften
+- Ondersteun interne overeenstemming over technologiedoelstellingen
+
+### Belangrijke aanbevelingen
+
+1. Voer grondig gebruikersonderzoek uit
+2. Begrijp gebruikersbehoeften volledig
+3. Stem technologische oplossingen af op gebruikerseisen
+
+## Gerelateerde hulpmiddelen
+
+- [Training gebruikersonderzoek](https://www.gebruikercentraal.nl/evenementen/)
+- [Digitale ondersteuning](https://www.digitaleoverheid.nl/overzicht-van-alle-onderwerpen/digitale-inclusie/mensen-helpen-om-met-digitalisering-om-te-gaan/)
+- [Handreiking Mobiele app Ontwikkeling en Beheer versie 4.0](https://www.noraonline.nl/images/noraonline/2/26/Handreiking_Mobiele_App_4.0.pdf) - Hoofdstuk 8 gebruikt "De gebruiker staat centraal" voor de ontwikkeling van mobiele apps.
+
+## Gerelateerde richtlijnen
+
+- [2. Zorg voor toegankelijkheid en inclusie](../toegankelijkheid/index.md)
+- [7. Maak privacy integraal](../privacy/index.md)
+- [14. Voldoe aan de servicestandaard](../servicestandaard/index.md)
