+++ docs/richtlijnen/integratie/index.md
@@ -1,0 +1,115 @@
+---
+title: "9. Integreer en pas technologie aan"
+summary: Je technologie moet werken met bestaande technologieën, processen en infrastructuur in je organisatie.
+relations:
+  - open-standaarden
+  - servicestandaard
+  - samenwerking
+---
+
+# 9. Integreer en pas technologie aan
+
+Nieuwe technologische oplossingen moeten naadloos integreren met bestaande systemen, processen en infrastructuur. Door zorgvuldige integratie verminder je complexiteit, verhoog je acceptatie en maximaliseer je de waarde van zowel nieuwe als bestaande investeringen.
+
+## Waarom is het belangrijk?
+
+- **Continuïteit**: Voorkom verstoring van kritieke diensten
+- **Kosteneffectiviteit**: Benut bestaande investeringen optimaal
+- **Gebruikservaring**: Creëer een samenhangende ervaring voor eindgebruikers
+- **Datastromen**: Zorg voor consistente en betrouwbare gegevensuitwisseling
+- **Complexiteitsreductie**: Verminder de beheerslast van gefragmenteerde systemen
+- **Risicovermindering**: Verlaag de kans op projectfalen door geleidelijke implementatie
+
+Door zorgvuldige integratie en aanpassing van technologie kun je een coherent en effectief digitaal landschap creëren dat waarde levert aan gebruikers en organisaties, terwijl het bestaande investeringen respecteert en toekomstige innovatie ondersteunt.
+
+## Hoe pas je het toe?
+
+<div class="direct-aan-de-slag">
+    <h3>Direct aan de slag</h3>
+
+    <div class="warning-banner" style="background-color: #fff3e0; padding: 0.5rem; border-left: 3px solid #ff9800; margin-bottom: 0.8rem;">
+        <strong>Work in Progress:</strong> De onderstaande functionaliteit is nog in ontwikkeling.
+    </div>
+
+    <div class="action-cards">
+        <div class="action-card">
+            <span class="wip-badge wip-badge-beschikbaar">beschikbaar</span>
+            <h4 >FSC</h4>
+            <p >Open source gegevensuitwisselingssysteem</p>
+            <a href="https://fsc-standaard.nl/" class="action-button" target="_blank">Verkennen</a>
+        </div>
+        <div class="action-card">
+            <span class="wip-badge wip-badge-beschikbaar">beschikbaar</span>
+            <h4 >Common Ground</h4>
+            <p >Visie voor moderne gegevensuitwisseling</p>
+            <a href="https://commonground.nl/" class="action-button" target="_blank">Leren</a>
+        </div>
+        <div class="action-card">
+            <span class="wip-badge wip-badge-beschikbaar">beschikbaar</span>
+            <h4 >Haven</h4>
+            <p >Platformonafhankelijke cloudhosting</p>
+            <a href="https://haven.commonground.nl/" class="action-button" target="_blank">Ontdekken</a>
+        </div>
+    </div>
+</div>
+
+### Integratie-aanpak
+
+1. **Ken je landschap**: Inventariseer bestaande systemen, processen en datastromen
+2. **Identificeer integratiepunten**: Bepaal waar nieuwe systemen moeten aansluiten
+3. **Evalueer impact**: Beoordeel de gevolgen van veranderingen op het geheel
+4. **Kies passende patronen**: Selecteer de juiste integratiemethoden en -technologieën
+5. **Plan gefaseerd**: Implementeer veranderingen geleidelijk om risico's te beperken
+6. **Test end-to-end**: Valideer de volledige integratie, niet alleen individuele componenten
+7. **Monitor prestaties**: Blijf integratiepunten bewaken op effectiviteit en efficiëntie
+
+### Integratiepatronen en -technologieën
+
+- **API's**: Application Programming Interfaces voor gestructureerde gegevensuitwisseling
+- **Middleware**: Software die verschillende applicaties verbindt
+- **Enterprise Service Bus (ESB)**: Centraal communicatiekanaal tussen systemen
+- **Microservices**: Modulaire architectuur die flexibele integratie ondersteunt
+- **ETL-processen**: Extract, Transform, Load voor databewegingen
+- **Event-driven architectuur**: Systemen communiceren via gebeurtenissen
+- **Hybride cloud**: Combinatie van on-premise en cloud-oplossingen
+
+### Uitdagingen en oplossingen
+
+- **Legacy systemen**: Ontwikkel adaptors of wrappers voor oudere technologie
+- **Dataformaten**: Gebruik standaarden en transformaties voor compatibiliteit
+- **Verschillende snelheden**: Implementeer bimodale IT om innovatie en stabiliteit te balanceren
+- **Beveiligingsgrenzen**: Ontwikkel veilige integratiepunten tussen verschillende beveiligingszones
+- **Organizational silos**: Bevorder samenwerking tussen afdelingen en teams
+
+### Best practices voor succesvolle integratie
+
+- **Gemeenschappelijke standaarden**: Bevorder het gebruik van open standaarden
+- **Loose coupling**: Minimaliseer directe afhankelijkheden tussen systemen
+- **Documentatie**: Houd architectuurdiagrammen en integratiepunten actueel
+- **Testen en monitoring**: Controleer regelmatig integratiepunten op werking
+- **Governance**: Ontwikkel richtlijnen voor het toevoegen van nieuwe systemen
+- **Change management**: Coördineer veranderingen over meerdere systemen heen
+
+### Overwegingen voor toekomstbestendigheid
+
+- Voorzie uitbreidbaarheid voor toekomstige integraties
+- Documenteer interfaces en datamodellen grondig
+- Ontwikkel strategieën voor migratie en uitfasering
+- Blijf open standaarden volgen voor langetermijncompatibiliteit
+- Evalueer regelmatig het technologielandschap op efficiëntie en effectiviteit
+
+## Gerelateerde standaarden
+
+- **[Common Ground](https://commonground.nl/)**: Een nieuwe informatiekundige visie voor moderne gegevensuitwisseling binnen de overheid. Common Ground werkt met herbruikbare componenten en het richtlijn van 'data bij de bron' om silovorming tegen te gaan en flexibele, modulaire IT-oplossingen te realiseren.
+
+- **[Federatieve Service Connectiviteit (FSC)](https://fsc-standaard.nl/)**: Een standaard voor een API-gateway met federatieve authenticatie. FSC beveiligt de verbindingen tussen partijen via wederzijdse digitale contracten en maakt veilige, gecontroleerde data-uitwisseling mogelijk.
+
+- **[Digikoppeling](https://www.logius.nl/diensten/digikoppeling)**: De standaard voor elektronisch berichtenverkeer tussen overheidsorganisaties. Digikoppeling beschrijft hoe gegevens veilig kunnen worden uitgewisseld, ongeacht het achterliggende systeem.
+
+- **[Haven](https://haven.commonground.nl/)**: Een standaard voor platformonafhankelijke cloudhosting, speciaal ontworpen voor het gemakkelijk herbruikbaar maken van applicaties tussen overheidsorganisaties.
+
+## Gerelateerde richtlijnen
+
+- [4. Gebruik open standaarden](../open-standaarden/index.md)
+- [8. Deel, hergebruik en werk samen](../samenwerking/index.md)
+- [14. Voldoe aan de servicestandaard](../servicestandaard/index.md)
