+++ docs/richtlijnen/privacy/index.md
@@ -1,0 +1,137 @@
+---
+title: "7. Maak privacy integraal"
+summary: Zorg ervoor dat gebruikersrechten worden beschermd door privacy te integreren als een essentieel onderdeel van je systeem.
+relations:
+  - veiligheid
+  - data
+  - algoritmen
+  - gebruikersbehoeften
+---
+
+# 7. Maak privacy integraal
+
+Privacy is een fundamenteel recht dat vanaf het begin moet worden ingebouwd in digitale systemen. Door privacy als kerncomponent te integreren, bescherm je niet alleen de rechten van gebruikers, maar bouw je ook vertrouwen op in overheidsdiensten.
+
+## Waarom is het belangrijk?
+
+- **Wettelijke verplichting**: Voldoe aan de Algemene Verordening Gegevensbescherming (AVG) en andere privacywetgeving
+- **Vertrouwen opbouwen**: Burgers vertrouwen systemen die hun privacy respecteren
+- **Risicovermindering**: Voorkom datalekken en privacyschendingen (zie [Maak veilige systemen](../veiligheid/index.md))
+- **Ethische verantwoordelijkheid**: Respecteer de rechten en waardigheid van individuen
+- **Goede dienstverlening**: Privacy-vriendelijke systemen zijn vaak gebruiksvriendelijker (zie [Voldoe aan de servicestandaard](../servicestandaard/index.md))
+- **Kosteneffectiviteit**: Voorkom dure boetes en herstelkosten door privacy vanaf het begin in te bouwen
+- **Datakwaliteit**: Privacy by design leidt vaak tot betere, meer nauwkeurige gegevensverzameling
+
+Door privacy integraal te maken in het ontwerpproces van digitale systemen, voldoe je niet alleen aan wettelijke vereisten, maar bouw je ook systemen die burgers vertrouwen en die hun rechten respecteren.
+
+## Hoe pas je het toe?
+
+<div class="direct-aan-de-slag">
+    <h3>Direct aan de slag</h3>
+
+    <div class="warning-banner" style="background-color: #fff3e0; padding: 0.5rem; border-left: 3px solid #ff9800; margin-bottom: 0.8rem;">
+        <strong>Work in Progress:</strong> De onderstaande functionaliteit is nog in ontwikkeling.
+    </div>
+
+    <div class="action-cards">
+        <div class="action-card">
+            <span class="wip-badge wip-badge-beschikbaar">beschikbaar</span>
+            <h4 >DPIA-model</h4>
+            <p >Model voor gegevensbeschermingseffectbeoordeling</p>
+            <a href="https://www.autoriteitpersoonsgegevens.nl/nl/onderwerpen/avg-europese-privacywetgeving/data-protection-impact-assessment-dpia" class="action-button" target="_blank">Downloaden</a>
+        </div>
+        <div class="action-card">
+            <span class="wip-badge wip-badge-beschikbaar">beschikbaar</span>
+            <h4 >Pre-scan DPIA & DPIA Formulier</h4>
+            <p >Online formulier voor gegevensbeschermingseffectbeoordeling</p>
+            <a href="https://minbzk.github.io/par-dpia-form/" class="action-button" target="_blank">Starten</a>
+        </div>
+        <div class="action-card">
+            <span class="wip-badge wip-badge-ontwikkeling">ontwikkeling</span>
+            <h4 >NICPET Tools</h4>
+            <p >Privacy-bevorderende technologieën</p>
+            <a href="https://nicpet.pleio.nl/" class="action-button" target="_blank">Verkennen</a>
+        </div>
+        <div class="action-card">
+            <span class="wip-badge wip-badge-concept">concept</span>
+            <h4 >Privacy Check</h4>
+            <p >Geautomatiseerde privacycheck</p>
+            <button disabled class="action-button">Starten</button>
+        </div>
+    </div>
+</div>
+
+### Privacy by Design vereisten
+
+- Integreer privacybescherming vanaf het begin van het ontwerpproces
+- Beschouw privacy als een essentieel onderdeel van systeemontwikkeling
+- Wees proactief in het verminderen van risico's op gegevensdiefstal
+
+### AVG-nalevingsoverwegingen
+
+Burgers hebben rechten op:
+
+- Toestemming voor gebruik van persoonlijke gegevens
+- Verwijdering van persoonlijke gegevens
+- Beperking van gegevensverwerking
+- Toegang tot en verplaatsing van persoonlijke gegevens
+
+### Aanbevolen impactbeoordelingsvragen
+
+Evalueer je project door te vragen:
+
+- Verzamel je nieuwe persoonlijke informatie?
+- Worden gegevens gedeeld met nieuwe partijen?
+- Gebruik je technologie die als privacy-indringend kan worden ervaren?
+- Kunnen beslissingen aanzienlijke impact hebben op individuen?
+- Bevatten de gegevens gevoelige persoonlijke details?
+
+### Kernrichtlijnen van Privacy by Design
+
+1. **Proactief, niet reactief**: Voorkom privacyproblemen voordat ze ontstaan
+2. **Privacy als standaardinstelling**: Maximale privacy zonder dat gebruikers actie hoeven te ondernemen
+3. **Privacy ingebouwd in het ontwerp**: Integreer privacy in de architectuur, niet als toevoeging
+4. **Volledige functionaliteit**: Vermijd valse tegenstellingen zoals privacy vs. veiligheid
+5. **End-to-end beveiliging**: Bescherm gegevens gedurende de hele levenscyclus
+6. **Zichtbaarheid en transparantie**: Houd praktijken open en verantwoordelijk
+7. **Respect voor gebruikersprivacy**: Houd de gebruiker centraal, bied controle en keuzemogelijkheden
+
+### Voordelen van Privacy by Design
+
+- Identificeer privacyproblemen vroeg
+- Verminder potentiële risico's
+- Vergroot organisatorisch privacybewustzijn
+- Zorg voor naleving van regelgeving
+
+### Praktische implementatiestappen
+
+- **Data-inventarisatie**: Identificeer welke persoonsgegevens worden verzameld en verwerkt
+- **Data minimalisatie**: Verzamel alleen gegevens die noodzakelijk zijn
+- **Privacyeffectbeoordeling (PIA)**: Voer deze uit bij nieuwe projecten of belangrijke wijzigingen
+- **Privacyverklaringen**: Maak deze duidelijk, toegankelijk en begrijpelijk
+- **Toestemming**: Zorg voor geldige toestemming waar nodig
+- **Bewaartermijnen**: Definieer en implementeer beleid voor gegevensretentie
+- **Rechten van betrokkenen**: Maak uitoefening van rechten (inzage, correctie, verwijdering) mogelijk
+- **Beveiligingsmaatregelen**: Implementeer passende technische en organisatorische maatregelen
+
+### Privacybevorderende technologieën
+
+- **Encryptie**: Bescherm gegevens in rust en tijdens transport
+- **Pseudonimisering**: Verwijder direct identificerende elementen
+- **Anonimisering**: Maak data niet-herleidbaar waar mogelijk
+- **Toegangscontrole**: Implementeer het richtlijn van minste privilege
+- **Privacymonitoring**: Detecteer en reageer op privacyproblemen
+
+## Gerelateerde hulpmiddelen
+
+- Raadpleeg Autoriteit Persoonsgegevens-richtlijnen
+- Bekijk de AVG (GDPR) regelgeving
+- Gebruik Data Protection Impact Assessment (DPIA) templates
+- [Handreiking Mobiele app Ontwikkeling en Beheer versie 4.0](https://www.noraonline.nl/images/noraonline/2/26/Handreiking_Mobiele_App_4.0.pdf) - Hoofdstuk 3.2 gebruikt de privacy by design richtlijnen voor de ontwikkeling van een mobiele app.
+
+## Gerelateerde richtlijnen
+
+- [1. Stel gebruikersbehoeften vast](../gebruikersbehoeften/index.md)
+- [6. Maak veilige systemen](../veiligheid/index.md)
+- [10. Maak beter gebruik van data](../data/index.md)
+- [11. Pas algoritmen verantwoord toe](../algoritmen/index.md)
