+++ mkdocs.yml
@@ -32,22 +32,22 @@ theme:
 
 nav:
   - Home: index.md
-  - Principes:
-      - Overzicht: principes/index.md
-      - 1. Gebruikersbehoeften: principes/gebruikersbehoeften/index.md
-      - 2. Toegankelijkheid: principes/toegankelijkheid/index.md
-      - 3. Open Source: principes/open-source/index.md
-      - 4. Open Standaarden: principes/open-standaarden/index.md
-      - 5. Cloud: principes/cloud/index.md
-      - 6. Veiligheid: principes/veiligheid/index.md
-      - 7. Privacy: principes/privacy/index.md
-      - 8. Samenwerking: principes/samenwerking/index.md
-      - 9. Integratie: principes/integratie/index.md
-      - 10. Data: principes/data/index.md
-      - 11. Algoritmen: principes/algoritmen/index.md
-      - 12. Inkoopstrategie: principes/inkoop/index.md
-      - 13. Duurzaamheid: principes/duurzaamheid/index.md
-      - 14. Servicestandaard: principes/servicestandaard/index.md
+  - Richtlijnen:
+      - Overzicht: richtlijnen/index.md
+      - 1. Gebruikersbehoeften: richtlijnen/gebruikersbehoeften/index.md
+      - 2. Toegankelijkheid: richtlijnen/toegankelijkheid/index.md
+      - 3. Open Source: richtlijnen/open-source/index.md
+      - 4. Open Standaarden: richtlijnen/open-standaarden/index.md
+      - 5. Cloud: richtlijnen/cloud/index.md
+      - 6. Veiligheid: richtlijnen/veiligheid/index.md
+      - 7. Privacy: richtlijnen/privacy/index.md
+      - 8. Samenwerking: richtlijnen/samenwerking/index.md
+      - 9. Integratie: richtlijnen/integratie/index.md
+      - 10. Data: richtlijnen/data/index.md
+      - 11. Algoritmen: richtlijnen/algoritmen/index.md
+      - 12. Inkoopstrategie: richtlijnen/inkoop/index.md
+      - 13. Duurzaamheid: richtlijnen/duurzaamheid/index.md
+      - 14. Servicestandaard: richtlijnen/servicestandaard/index.md
   - Over NeRDS:
       - Over-NeRDS/index.md
       - Bijdragen: Over-NeRDS/CONTRIBUTING.md
